requirements.txt
python-telegram-bot==13.15
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
flask==2.3.3